+++ packages/blocks-core/js/libs/wordpress/quote/test/functioanlity.blocks-2.e2e.cy.js
@@ -41,6 +41,18 @@ describe('Quote Block', () => {
 			'core/paragraph',
 		]);
 
+		cy.checkBlockStatesPickerItems([
+			'elements/citation',
+			'core/paragraph',
+			'elements/link',
+			'elements/bold',
+			'elements/italic',
+			'elements/kbd',
+			'elements/code',
+			'elements/span',
+			'elements/mark',
+		]);
+
 		//
 		// 1. Edit Block
 		//