+++ packages/blocks-core/CHANGELOG.md
@@ -40,6 +40,7 @@
 - `Post Navigation Link` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Post Template` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Post Title` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
+- `Quote` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 - `Pullquote` block: `strong`, `italic`, `kbd`, `span`, `code` and `mark` inner blocks added.
 
 
@@ -82,6 +83,7 @@
 - Added E2E tests to check functionality `Post Navigation Link` new inner block.
 - Added E2E tests to check functionality `Post Template` new inner block.
 - Added E2E tests to check functionality `Post Title` new inner block.
+- Added E2E tests to check functionality `Quote` new inner block.
 - Added E2E tests to check functionality `Pullquote` new inner block.
 
 ## 1.4.0 (2025-05-19)